+++ iconic_api/models/__init__.py
@@ -1,4 +1,5 @@
 from .openapi_generated import *
+from .order_search import *
 from .api_requests import *
 from .attribute import *
 from .stock import *